feat: 댓글등록 엔터이벤트 추가, 댓글 삭제 시 확인창

+++ src/component/postdetail/comment/deletecommentbutton.tsx
@@ -8,23 +8,27 @@ interface DeleteCommentButtonProps {
 }
 
 const DeleteCommentButton: React.FC<DeleteCommentButtonProps> = ({ token, id, onCommentUpdate }) => {
-  const handleDeleteClick = async () => {try {
+  const handleDeleteClick = async () => {
+    const confirmDelete = window.confirm("댓글을 삭제하시겠습니까?");
+  if(confirmDelete){
+    try {
       const response = await deleteComment(token, id);
       if (response === 200) {
-        alert("삭제 성공");
+        // alert("삭제 성공");
         onCommentUpdate();
       }
       else if (response === 403) {
         alert('이 댓글의 삭제에 대한 권한이 없습니다.');
       }
       else {
-        alert('삭제 에러');
+        // alert('삭제 에러');
       }
     }
     catch (error) {
      console.error("삭제 에러", error);
      alert('삭제 에러');
     }
+  }
   };
 
   return <button onClick={handleDeleteClick}>삭제</button>;
